Working Programs
Due to "duplicate errors", I disabled the colorSensorTest program and
androidStudioDemo program so that other programs can still be run. I
also wrote a working clawTest program and combined it with the
teleopTrial program to make teleopDrivetrainArm. This program uses two
controllers, one for the arm and the other for drivetrain.